app/requirements.txt
aniso8601==9.0.1
click==8.1.2
colorama==0.4.4
Flask==2.1.1
Flask-JSON==0.3.4
Flask-RESTful==0.3.9
flask-restful-swagger-2==0.35
gunicorn==20.1.0
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
neo4j-driver==4.3.6
neobolt==1.7.17
neomodel==4.0.8
numpy==1.21.6
pandas==1.3.5
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
Shapely==1.7.1
six==1.16.0
typing_extensions==4.2.0
Werkzeug==2.1.1
zipp==3.8.0